+++ allennlp/common/util.py
@@ -328,7 +328,7 @@ def push_python_path(path: PathType) -> ContextManagerFunctionReturnType[None]:
         sys.path.remove(path)
 
 
-def import_module_and_submodules(package_name: str) -> None:
+def import_module_and_submodules(package_name: str, exclude: Optional[List[str]] = None) -> None:
     """
     Import all submodules under the given package.
     Primarily useful so that people using AllenNLP as a library
@@ -342,7 +342,8 @@ def import_module_and_submodules(package_name: str) -> None:
     # the end won't hurt anything.
     with push_python_path("."):
         # Import at top level
-        module = importlib.import_module(package_name)
+        if not exclude or package_name not in exlude:
+            module = importlib.import_module(package_name)
         path = getattr(module, "__path__", [])
         path_string = "" if not path else path[0]
 
@@ -353,7 +354,7 @@ def import_module_and_submodules(package_name: str) -> None:
             if path_string and module_finder.path != path_string:  # type: ignore[union-attr]
                 continue
             subpackage = f"{package_name}.{name}"
-            import_module_and_submodules(subpackage)
+            import_module_and_submodules(subpackage, exclude=exclude)
 
 
 def peak_cpu_memory() -> Dict[int, int]: